`fixQuad` is no longer used by production code
